Checkout revs to [repo]-cache
When https://gerrit.wikimedia.org/r/#/c/241684/ merges to preserve
backwards compatibility with trebuchet, this will need to merge.

This makes the final path to the server executable
/srv/deployment/mockbase/deploy/mockbase/server

It also puts in place the /srv/deployment/mockbase/deploy-cache
directory which is the directory where revs, cache, and current will be
stored.